+++ CHANGES.md
@@ -1,8 +1,12 @@
-## 0.0.12 (2023/05/18)
+## 0.0.14 (2023/11/02)
+* FIXED: multiple bugs in AliveKeeper
+
+## 0.0.13 (2023/05/18)
 * ADDED: maximum concurrency setting
 * ADDED: concurrency policies for AliveKeeper
 * FIXED: safe sync policy was deadlocking in some cases
 * ADDED: ability for agents to end their work without being disposed
+* ADDED: cancellation token can be passed to agents
 
 ## 0.0.8 (2023/05/06)
 * CHANGED: build process